Moved helper functions to utils.  Moved Crosshair data fetching to slice
