+++ src/app/utils.ts
@@ -275,3 +275,30 @@ export function calculateTotalFees(order: any): number {
     ? roundTo(totalFees, 4, RoundType.NEAREST)
     : totalFees;
 }
+
+//Chart Helper Functions
+export const formatPercentageChange = (percChange: number | null): string => {
+  if (percChange !== null) {
+    return `(${percChange.toFixed(2)}%)`;
+  }
+  return "(0.00%)";
+};
+
+export const computePercentageChange = (
+  currentClose: number | null,
+  prevClose: number | null
+): number | null => {
+  if (currentClose !== null && prevClose !== null) {
+    const difference = currentClose - prevClose;
+    return (difference / prevClose) * 100;
+  }
+  return null;
+};
+
+export const getVolumeBarColor = (
+  currentClose: number,
+  prevClose: number
+): string => {
+  const percChange = ((currentClose - prevClose) / prevClose) * 100;
+  return percChange < 0 ? "#f44336" : "#4caf50";
+};
